Split CompilationUnit into CompilationUnit + SourceUnit
