+++ crytic_compile/platform/etherscan.py
@@ -76,13 +76,15 @@ def _handle_bytecode(crytic_compile: "CryticCompile", target: str, result_b: byt
 
     compilation_unit = CompilationUnit(crytic_compile, str(target))
 
-    compilation_unit.contracts_names.add(contract_name)
+    source_unit = compilation_unit.create_source_units(contract_filename)
+
+    source_unit.contracts_names.add(contract_name)
     compilation_unit.filename_to_contracts[contract_filename].add(contract_name)
-    compilation_unit.abis[contract_name] = {}
-    compilation_unit.bytecodes_init[contract_name] = bytecode
-    compilation_unit.bytecodes_runtime[contract_name] = ""
-    compilation_unit.srcmaps_init[contract_name] = []
-    compilation_unit.srcmaps_runtime[contract_name] = []
+    source_unit.abis[contract_name] = {}
+    source_unit.bytecodes_init[contract_name] = bytecode
+    source_unit.bytecodes_runtime[contract_name] = ""
+    source_unit.srcmaps_init[contract_name] = []
+    source_unit.srcmaps_runtime[contract_name] = []
 
     compilation_unit.compiler_version = CompilerVersion(
         compiler="unknown", version="", optimized=False
